requirements.txt
openai>=1.12.0
python-dotenv==1.0.1
pyyaml==6.0.1
markdownify==0.11.6
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
